Ingest old_entity table; check for presence of sqlite3 file to make development easier

+++ task/load.sh
@@ -12,8 +12,10 @@ DATABASE_NAME=${DATABASE%.*}
 
 URL="https://$S3_BUCKET.s3.eu-west-2.amazonaws.com/$S3_KEY"
 
-echo "downloading from $URL"
-curl -O $URL
+if ! [ -f $DATABASE_NAME.sqlite3 ]; then
+  echo "downloading from $URL"
+  curl -O $URL
+fi
 
 echo "exporting $DATABASE"
 sqlite3 $DATABASE ".read sql/export_$DATABASE_NAME.sql"
